workers/av2_datasets/requirements.txt
celery>=5.4.0,<6
kombu>=5.4.0,<6
redis>=5.0.0,<6
rich>=13.9.2,<14
pydantic>=2.9.2,<3
opencv-python>=4.10.0,<5
PyYAML>=6.0,<7
shapely>=2.0,<3
pillow>=11.0,<12
matplotlib>=3.9,<4